+++ packages/functions/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @tango-ts/functions
 
+## 0.9.0
+
 ## 0.8.1
 
 ### Patch Changes